+++ CsAutoAdvancedTraining/Repository/DAL/XmlConfigEntities/XmlConfig.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Xml.Serialization;
+
+namespace Repository.DAL.XmlConfigEntities
+{
+    [XmlRoot("config")]
+    public class XmlConfig 
+    {
+        [XmlIgnore]
+        public string ConfigName { get; set; }
+
+        [XmlElement("browser")]
+        public List<XmlBrowserConfig> BrowserConfigs { get; set;  }
+    }
+}
