+++ CsAutoAdvancedTraining/Repository/DAL/XmlConfigEntities/XmlTestStep.cs
@@ -1,0 +1,13 @@
+﻿using System.Xml.Serialization;
+
+namespace Repository.DAL.XmlConfigEntities
+{
+    public class XmlTestStep
+    {
+        [XmlAttribute("number")]
+        public int StepNumber { get; set; }
+
+        [XmlText]
+        public string StepText { get; set; }
+    }
+}
