+++ CsAutoAdvancedTraining/Repository/Enums/Browser.cs
@@ -1,0 +1,9 @@
+﻿namespace Repository.Enums
+{
+    public enum Browser
+    {
+        Chrome,
+        Firefox,
+        Edge
+    }
+}
