+++ CsAutoAdvancedTraining/Repository/Enums/TestConfigType.cs
@@ -1,0 +1,8 @@
+﻿namespace Repository.Enums
+{
+    public enum TestConfigType
+    {
+        Xml, 
+        Json
+    }
+}
