+++ CsAutoAdvancedTraining/Repository/Enums/UserRole.cs
@@ -1,0 +1,9 @@
+﻿namespace Repository.Enums
+{
+    public enum UserRole
+    {
+        Admin,
+        Editor,
+        Guest
+    }
+}
