+++ CsAutoAdvancedTraining/Repository/Interfaces/IRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace Repository.Interfaces
+{
+    public interface IRepository : IRepositoryReader, IRepositoryWriter
+    {
+    }
+}
