+++ CsAutoAdvancedTraining/Repository/Interfaces/IRepositoryReader.cs
@@ -1,0 +1,9 @@
+﻿using Repository.Models;
+
+namespace Repository.Interfaces
+{
+    public interface IRepositoryReader
+    {
+        Config GetConfig();
+    }
+}
