+++ CsAutoAdvancedTraining/Repository/Interfaces/IRepositoryWriter.cs
@@ -1,0 +1,9 @@
+﻿using Repository.Models;
+
+namespace Repository.Interfaces
+{
+    public interface IRepositoryWriter
+    {
+        void WriteConfig(Config config);
+    }
+}
