+++ CsAutoAdvancedTraining/Repository/Models/TestStep.cs
@@ -1,0 +1,9 @@
+﻿namespace Repository.Models
+{
+    public class TestStep
+    {
+        public int StepNumber { get; set; }
+
+        public string StepText { get; set; }
+    }
+}
