+++ CsAutoAdvancedTraining/Repository/Models/User.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+using Repository.Enums;
+
+namespace Repository.Models
+{
+    public class User
+    {
+        public UserRole Role { get; set; }
+
+        public string Login { get; set; }
+
+        public string Password { get; set; }
+
+        public List<Test> Tests { get; set; }
+    }
+}
