+++ CsAutoAdvancedTraining/Repository/Repository.csproj
@@ -4,17 +4,16 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{DA0A818C-03D6-4C75-A3FF-71517DFDA22F}</ProjectGuid>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>TestConfiguration</RootNamespace>
-    <AssemblyName>TestConfiguration</AssemblyName>
+    <ProjectGuid>{01249789-94D6-4FBF-B87D-0C1908C9FB1B}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>Repository</RootNamespace>
+    <AssemblyName>Repository</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <Deterministic>true</Deterministic>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
@@ -24,7 +23,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>bin\Release\</OutputPath>
@@ -34,9 +32,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
-    <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Memory, Version=4.0.1.2, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL" />
     <Reference Include="System.Text.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -46,12 +42,25 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DAL\JsonConfigEntities\JsonConfig.cs" />
+    <Compile Include="DAL\XmlConfigEntities\XmlBrowserConfig.cs" />
+    <Compile Include="DAL\XmlConfigEntities\XmlConfig.cs" />
+    <Compile Include="DAL\XmlConfigEntities\XmlTest.cs" />
+    <Compile Include="DAL\XmlConfigEntities\XmlTestStep.cs" />
+    <Compile Include="DAL\XmlConfigEntities\XmlUser.cs" />
+    <Compile Include="Enums\Browser.cs" />
+    <Compile Include="Enums\TestConfigType.cs" />
+    <Compile Include="Enums\UserRole.cs" />
     <Compile Include="Extensions\ConfigExtensions.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="Interfaces\IRepository.cs" />
+    <Compile Include="Interfaces\IRepositoryReader.cs" />
+    <Compile Include="Interfaces\IRepositoryWriter.cs" />
+    <Compile Include="Models\BrowserConfig.cs" />
+    <Compile Include="Models\Config.cs" />
+    <Compile Include="Models\Test.cs" />
+    <Compile Include="Models\TestStep.cs" />
+    <Compile Include="Models\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="App.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
