+++ CsAutoAdvancedTraining/TestConfiguration/App.config
@@ -6,6 +6,14 @@
   <appSettings>
     <add key="inputPath" value="C:\TestConfigRepository\Input" />
     <add key="outputPath" value="C:\TestConfigRepository\Output" />
+    
+    <add key="xmlRepositoryAssemblyPath" value="C:\Users\IlyaOrlenko\source\repos\CsAutoAdvancedTraining\CsAutoAdvancedTraining\XmlRepository\bin\Debug\XmlRepository.dll"/>
+    <add key="jsonRepositoryAssemblyPath" value="C:\Users\IlyaOrlenko\source\repos\CsAutoAdvancedTraining\CsAutoAdvancedTraining\JsonRepository\bin\Debug\JsonRepository.dll"/>
+
+    <add key="xmlRepositoryClassName" value="Xml.XmlRepository"/>
+    <add key="jsonRepositoryClassName" value="Json.JsonRepository"/>
+    <add key="getConfigMethodName" value="GetConfig"/>
+    <add key="writeConfigMethodName" value="WriteConfig"/>
   </appSettings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
