--- CsAutoAdvancedTraining/TestConfiguration/ConfigFiles/cfg.json
@@ -1,284 +1,0 @@
-{
-	"Browsers": [
-		{
-			"BrowserName": "Chrome",
-			"BrowserVersion": "107",
-			"Users": [
-				{
-					"Login": "guestLogin",
-					"Password": "ytrewq",
-					"Role": "Guest",
-					"Tests": [
-						{
-							"ExpectedResult": "Support answeres asap",
-							"Id": 1,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Fill out the feedback form"
-								},
-								{
-									"StepNumber": 3,
-									"StepText": "Forget to enter your contact email"
-								},
-								{
-									"StepNumber": 4,
-									"StepText": "Click submit"
-								},
-								{
-									"StepNumber": 5,
-									"StepText": "Wait forever"
-								}
-							],
-							"Title": "Chrome Test 01 - Guest"
-						}
-					]
-				},
-				{
-					"Login": "user1",
-					"Password": "1qaz!QAZ",
-					"Role": "Guest",
-					"Tests": [
-						{
-							"ExpectedResult": "App doesn't crash",
-							"Id": 2,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Crack there"
-								},
-								{
-									"StepNumber": 3,
-									"StepText": "Overload there"
-								},
-								{
-									"StepNumber": 4,
-									"StepText": "Monkey mode there"
-								}
-							],
-							"Title": "Chrome Test 02 - Guest"
-						},
-						{
-							"ExpectedResult": "App doesn't crash",
-							"Id": 3,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Crack there"
-								},
-								{
-									"StepNumber": 3,
-									"StepText": "Overload there"
-								},
-								{
-									"StepNumber": 4,
-									"StepText": "Monkey mode there"
-								}
-							],
-							"Title": "Chrome Test 03 - Guest"
-						}
-					]
-				},
-				{
-					"Login": "administrator",
-					"Password": "1qaz!QAZ",
-					"Role": "Admin",
-					"Tests": [
-						{
-							"ExpectedResult": "Everything works fine",
-							"Id": 4,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Click there"
-								}
-							],
-							"Title": "Chrome Test 04 - Admin"
-						},
-						{
-							"ExpectedResult": "Everything works extremely fine",
-							"Id": 5,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go hard there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Click hard there"
-								}
-							],
-							"Title": "Chrome Test 05 - Admin"
-						}
-					]
-				}
-			]
-		},
-		{
-			"BrowserName": "Edge",
-			"BrowserVersion": "105",
-			"Users": [
-				{
-					"Login": "administrator",
-					"Password": "2wsx@WSX",
-					"Role": "Admin",
-					"Tests": [
-						{
-							"ExpectedResult": "Bla bla test 1",
-							"Id": 8,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Bla 1-1-1"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Bla 1-1-2"
-								}
-							],
-							"Title": "Edge Test 01 - Admin 1"
-						},
-						{
-							"ExpectedResult": "Bla bla test 2",
-							"Id": 9,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Bla 1-2-1"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Bla 1-2-2"
-								}
-							],
-							"Title": "Edge Test 02 - Admin 1"
-						}
-					]
-				},
-				{
-					"Login": "administrator",
-					"Password": "",
-					"Role": "Admin",
-					"Tests": [
-						{
-							"ExpectedResult": "Bla bla test 1",
-							"Id": 10,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Bla 2-1-1"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Bla 2-1-2"
-								}
-							],
-							"Title": "Edge Test 03 - Admin 2"
-						},
-						{
-							"ExpectedResult": "Bla bla test 2",
-							"Id": 11,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Bla 2-2-1"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Bla 2-2-2"
-								}
-							],
-							"Title": "Edge Test 04 - Admin 2"
-						}
-					]
-				}
-			]
-		},
-		{
-			"BrowserName": "Firefox",
-			"BrowserVersion": "106",
-			"Users": [
-				{
-					"Login": "guestLogin",
-					"Password": null,
-					"Role": "Guest",
-					"Tests": [
-						{
-							"ExpectedResult": "Support answeres ASAP",
-							"Id": 6,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Fill out the feedback form"
-								},
-								{
-									"StepNumber": 3,
-									"StepText": "Forget to enter your contact email"
-								},
-								{
-									"StepNumber": 4,
-									"StepText": "Click submit"
-								},
-								{
-									"StepNumber": 5,
-									"StepText": "Wait forever"
-								}
-							],
-							"Title": "Firefox Test 01 - Guest"
-						}
-					]
-				},
-				{
-					"Login": "user2",
-					"Password": null,
-					"Role": "Editor",
-					"Tests": [
-						{
-							"ExpectedResult": "App doesn't crash",
-							"Id": 7,
-							"Steps": [
-								{
-									"StepNumber": 1,
-									"StepText": "Go there"
-								},
-								{
-									"StepNumber": 2,
-									"StepText": "Crack there"
-								},
-								{
-									"StepNumber": 3,
-									"StepText": "Overload there"
-								},
-								{
-									"StepNumber": 4,
-									"StepText": "Monkey mode there"
-								}
-							],
-							"Title": "Firefox Test 02 - Editor"
-						}
-					]
-				}
-			]
-		}
-	]
-}
