--- CsAutoAdvancedTraining/TestConfiguration/DAL/XmlConfigEntities/XmlConfig.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Xml.Serialization;
-
-namespace TestConfiguration.DAL
-{
-    [XmlRoot("config")]
-    public class XmlConfig 
-    {
-        [XmlIgnore]
-        public string ConfigName { get; set; }
-
-        [XmlElement("browser")]
-        public List<XmlBrowserConfig> BrowserConfigs { get; set;  }
-    }
-}
