--- CsAutoAdvancedTraining/TestConfiguration/DAL/XmlConfigEntities/XmlTestStep.cs
@@ -1,13 +1,0 @@
-﻿using System.Xml.Serialization;
-
-namespace TestConfiguration.DAL
-{
-    public class XmlTestStep
-    {
-        [XmlAttribute("number")]
-        public int StepNumber { get; set; }
-
-        [XmlText]
-        public string StepText { get; set; }
-    }
-}
