--- CsAutoAdvancedTraining/TestConfiguration/Enums/Browser.cs
@@ -1,9 +1,0 @@
-﻿namespace TestConfiguration.Enums
-{
-    public enum Browser
-    {
-        Chrome,
-        Firefox,
-        Edge
-    }
-}
