--- CsAutoAdvancedTraining/TestConfiguration/Enums/TestConfigType.cs
@@ -1,8 +1,0 @@
-﻿namespace TestConfiguration.Enums
-{
-    public enum TestConfigType
-    {
-        Xml, 
-        Json
-    }
-}
