--- CsAutoAdvancedTraining/TestConfiguration/Enums/UserRole.cs
@@ -1,9 +1,0 @@
-﻿namespace TestConfiguration.Enums
-{
-    public enum UserRole
-    {
-        Admin,
-        Editor,
-        Guest
-    }
-}
