--- CsAutoAdvancedTraining/TestConfiguration/Interfaces/IRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace TestConfiguration.Interfaces
-{
-    internal interface IRepository : IRepositoryReader, IRepositoryWriter
-    {
-    }
-}
