--- CsAutoAdvancedTraining/TestConfiguration/Interfaces/IRepositoryReader.cs
@@ -1,9 +1,0 @@
-﻿using TestConfiguration.Models;
-
-namespace TestConfiguration.Interfaces
-{
-    internal interface IRepositoryReader
-    {
-        Config GetConfig();
-    }
-}
