--- CsAutoAdvancedTraining/TestConfiguration/Interfaces/IRepositoryWriter.cs
@@ -1,9 +1,0 @@
-﻿using TestConfiguration.Models;
-
-namespace TestConfiguration.Interfaces
-{
-    internal interface IRepositoryWriter
-    {
-        void WriteConfig(Config config);
-    }
-}
