--- CsAutoAdvancedTraining/TestConfiguration/Models/Config.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace TestConfiguration.Models
-{
-    public class Config
-    {
-        public string ConfigName { get; set; }
-
-        public List<BrowserConfig> BrowserConfigs { get; set;  }
-    }
-}
