--- CsAutoAdvancedTraining/TestConfiguration/Models/TestStep.cs
@@ -1,9 +1,0 @@
-﻿namespace TestConfiguration.Models
-{
-    public class TestStep
-    {
-        public int StepNumber { get; set; }
-
-        public string StepText { get; set; }
-    }
-}
